Move integration tests into their own project
