+++ src/ApiService/IntegrationTests/DownloadTests.cs
@@ -2,14 +2,13 @@
 using System;
 using System.Net;
 using System.Net.Http;
+using IntegrationTests.Fakes;
 using Microsoft.OneFuzz.Service;
-using Tests.Fakes;
 using Xunit;
 using Xunit.Abstractions;
-
 using Async = System.Threading.Tasks;
 
-namespace Tests.Functions;
+namespace IntegrationTests;
 
 [Trait("Category", "Integration")]
 public class AzureStorageDownloadTest : DownloadTestBase {
