+++ src/ApiService/IntegrationTests/InfoTests.cs
@@ -1,14 +1,13 @@
 ﻿
 using System;
 using System.Net;
+using IntegrationTests.Fakes;
 using Microsoft.OneFuzz.Service;
-using Tests.Fakes;
 using Xunit;
 using Xunit.Abstractions;
-
 using Async = System.Threading.Tasks;
 
-namespace Tests.Functions;
+namespace IntegrationTests;
 
 [Trait("Category", "Integration")]
 public class AzureStorageInfoTest : InfoTestBase {
