+++ src/ApiService/IntegrationTests/_FunctionTestBase.cs
@@ -5,13 +5,13 @@ using ApiService.OneFuzzLib.Orm;
 using Azure.Data.Tables;
 using Azure.Storage;
 using Azure.Storage.Blobs;
+using IntegrationTests.Fakes;
 using Microsoft.Azure.Functions.Worker.Http;
 using Microsoft.OneFuzz.Service;
 using Microsoft.OneFuzz.Service.OneFuzzLib.Orm;
-using Tests.Fakes;
 using Xunit.Abstractions;
 
-namespace Tests.Functions;
+namespace IntegrationTests;
 
 // FunctionTestBase contains shared implementations for running
 // functions against live Azure Storage or the Azurite emulator.
